Removed W_CRCNumber from the public API as it is of no use to plugins. The rest of the API allows access to our internal file system using an abstract interface which hides the detail of whether resources have been loaded from Zip, the real local file system or indeed the WAD lump directory. Simply taking a CRC of the IWAD's lump directory is of no benefit for writing to save games (the only possible reason I can see why this was made public).
Removed W_OpenAuxiliary from the public API. We abstract access to WAD lumps so the game should never need to call it. If in-WAD data must be read on game-side after initial game load - then it should use the abstract file loading interface instead.

Either we abstract file/lump access or we don't. The indecisive nature of our existing API is very confusing for a third-party to understand (been there myself). Somewhat more importantly, however, is that the mere presence of these lower-level accessors have a very negative impact on the development of engine-internal subsystems, as we must retain API backward compatibility.