Server sends player's initial spawn position
When the map changes, it is possible that it takes a while
before the entire world state has been transferred to the
client. During this time, the client may be unaware of where
the player's mobj is supposed to be located. Instead of
the client showing a blank screen, the server sends the
initial spawn position as soon as it has been decided, so
that the client can move its local mobj accordingly, and
wait for the rest of the world state to catch up.

However, it seems that currently the server is a bit too
careful when sending data: even on a fully local connection,
the amount of data is restricted so that the world state is
transferred at quite a slow pace. This will be addressed
when the network connection code is revised.