Began dividing up the menu system, separating the underlying (generic) UI widget subsystem from the game menu implementation.
So far so good, its a relatively clean break with only three
interdependencies left to resolve: sounds, widget animation
(i.e., lack of tickers) and the menu cursor positioning.

Also fixed a couple of bugs in the process:
* Menu overlay backgrounds (the dark fade when grabbing event
input for bindings) did not fill the window if the menu was not
being stretched.
* CVar-linked colorboxes were not updating their associated
variables upon leaving the color widget.

Next step: cleanup headers.