Added new public API routine R_GetPatchName for querying the name associated with a unique patch identifier.
I opted not to make use of patchinfo_t because the games currently
statically-allocate copies of this struct.

This needs cleaning up; either:
a) engine manages the extended info and stores it so the game can
retrieve an immutable pointer to it whenever it is required
b) game asks the engine to dynamically allocate the info which the
game should then take care to destroy when no longer needed.