Heretic: Fixing chicken morph
Located a few issues with the code (such as using __JHERETIC
instead of __JHERETIC__) that were messing things up.
Currently the client is getting the message about morphing,
but the server seems to switch the mobj to the wrong attack
state when the chicken attacks. I suppose the player class
has something to do with this.