Completed adaptation of the automap into a native UI widget (only Doom at this point. The other games will FTBFS presently).
There is still much more I'd like to do here (e.g., extract the automap
mask functionality and make it a general widget feature). However I don't
want to get bogged down with the implementation of a particular class of
widget and lose the momentum needed to complete the UI subsystem refactor.
Therefore I've decided enough is enough and drawn the line here. We can
always return to this at a later time.

Now to update the other games...