Refactor: Removed all arguments from FR_DrawText that are now in the font renderer attribute stack. Refactored callers to configure attribute state.
Fixed: Map name displayed while the automap is open is now drawn with no
additional effects, like the rest of the hud.

Changed: Important game messages (e.g., those produced when attempting to
quit the game) are now considered to be part of the menu subsystem. These
messages are now drawn with the same attributes as the rest of the menu.