R_SetViewPortPlayer() replaces DD_DISPLAYPLAYER
The engine's displayPlayer was used internally to store
which player is being rendered in a particular viewport.
However, the game was setting the displayPlayer value
and expecting it to affect the console player.

DD_DISPLAYPLAYER is no longer writable through the
API. R_SetViewPortPlayer() must be used instead, as
it will set the viewed player for a specific player (in a
specific view port).

Fixes the problem of Hexen's view going blank after
disconnecting from a netgame and starting a new game.