Fixed: Clientside notification about moving planes
When the client was moving a plane, it never told the game about this.
Consequently, the game logic was not properly updated with the new
plane height.

There are still issues with remote objects on moving planes because
the planes are moving at different times on serverside and clientside.
The client needs to be a bit more clever when applying the Z
coordinate from a delta.