Added Smoother: calculating player movement paths
de::Smoother is a class whose purpose is to reproduce a smooth
movement path out of a handful of distinct points.

The current implementation uses a simple linear interpolation.
It is used on the server to reproduce the client movement,
and on the client to smooth out the movement of clmobjs
belonging to remote players.

The Smoother class is part of the public API.