Apply movement smoothing in the renderer
Rather than applying it in the game logic, the smoother
is now applied in the renderer when the mobj becomes
a vissprite.

Ideally it should be done in game logic, but I was having
problems applying it correctly with the current "hidden
clmobj / visible mobj" mechanism where certain
properties of a clmobj are copied to the visible mobj
when a delta is received. The smoother works better
when the updates are continuous.