Client: Fixed initial plane heights update
The client receives a number of plane height updates
when joining a game. However, the viewer is reset
immediately after a map is loaded, so the updates
were not carried out as they should've been. Now the
reset will explicitly update the visible height of the
planes so they match the logical heights.