Utilize the auxiliary lump directory to load and validate the list of required and identity-keyed resources in the IWADs, before any attempt to load a game.
Note: Due to the issue outlined in the notes of the previous commit, the logic
for this isn't quite correct just yet. This is not an issue presently, however,
because the order of DD_StartupWorker's subroutines will ensure that only the
auxiliary lump directory is populated/in-use at this time.