Client: Fixed Z coord and mobjinfo problems
If a clmobj is supposed to be on the floor, the client will
now figure out the floor Z height at the position and adjust
the clmobj's Z coordinate accordingly.

Also, the type and mobjinfo of clmobjs was updated
incorrectly, causing some mobjs to freeze on client side
(e.g., Heretic splashes).