Refactoring network messaging
All network messages are now written and read using
the engine's Writer and Reader classes.

Added the build option DENG_WRITER_TYPECHECK
for verifying that the correct type of values are read back
(only for debugging purposes, as it increases the size of
the messages by including the type markers).