Render a few black frames just before shutdown.
On some displays when running fullscreen, the previously rendered
frame remains visible during a video context switch if not playing
with the same mode as used on the desktop. This can result in rather
ugly scaling artefacts which can be quite jarring, or worse, it can
be considered indicative of a less-than-graceful shutdown.