Store tangent and bitangent vectors in (map) surface_t.
For the purposes of lighting an decoration projection, these vectors
are frequently needed and by not storing them, it means repeatedly
calculating them each time they are needed. These vectors are needed
often enough that storing them along with the normal makes sense.

Updated surface debug display to draw all tangent space vectors.

Project simple mobj shadows similar to dynlights

This approach to drawing the shadows can naturally result in rather
a lot of overdraw. The upside however is that there far fewer visible
artefacts than the older algorithm. Neither is ideal so my current
plan is to implement a heuristic analysis of a mobj's shadow-casting
properties to determine which algorithm to use in a given situation.
In practice this will be pretty simple; close and large shadows use
the accurate method, others use the simple method.

The old algorithm for mobj shadows has been disabled temporarily.