Fixed: Font renderer drawing of gitter fx
Gitter for a fragment of a multipart text block was drawn with along the
with each fragment rather than in a separate drawing pass. So any case
scaled text (such as that used in DOOM's game menu) would see the glitter
for the previous text fragment drawn over by the character patches for
the current fragment.

Clearly we need to start thinking about optimizing the font renderer...