Win32: Include msvcr100.dll in the distribution
As libpng and libcurl have been compiled with VC++ 10.0, they
are dependent on msvcr100.dll. This shared library is part of
the MSVC2010 redistributable files and is not found by default
on Windows installations.

The Windows build is now working on 64-bit Windows 7 (running
as 32-bit) as well as 32-bit Windows.