Fixed: Segfault in rendSegSection when using rend-tex 2
When using rend-tex2 there are two materials in use for any given
surface, the one which is linked to the surface being rendered in
the map data and the material used as a replacement in this mode.

There is also the possibility that the surface has no material
bound to it (e.g., if the texture/flat reference in the archived
copy of the map cannot be resolved to a known material). In which
case the surface material will be substituted for the "missing"
material which is always available.

The fault occurs when rend-tex 2 is used while a surface which has
no material bound to it comes into view and an attempt to dereference
NULL is made.

Todo: rendSegSection is dire need of a cleanup. Do this prior to
merging back to master.