gcc: Disabled most warnings
There are hundreds of warnings from the old C code, such as
signed/unsigned comparisons and comparisons that are always true
due to value range. These are now disabled as it is unlikely
that all of these will be fixed for the legacy code.

TODO: Enable strict warnings for new C++ code.