versions.pri: Read version numbers from header files
The build_version.py script was improved to allow reading
version number information for all the components directly
from header files.

When build_version.py is invoked from the command line,
it expects to be given a header file name as the only argument.
When imported as a module, it finds the Doomsday version
and exports it as global variables.

build_number.py was changed so that it only prints the build
number if given the --print argument. By default it can be
imported as a module and the todays_build() function returns
the build number as a string.