LegacyNetwork is done
The API was slightly modified so that waiting is not done by libdeng2.
Any waiting has to take into account that there is an event loop that
needs to run and waiting will block the current thread.

Also, messages are received in buffers allocated and freed by
LegacyNetwork. This is because deng2 sockets use a block-based
transfer protocol. It will be necessary to modify the network code
so that it does not initially use a plain TCP stream for
communications. This should be an improvement from a security
perspective.