Finished refactoring sys_network for LegacyNetwork
The low-level networking code is now completely dependent on
the libdeng2-based network code. The initial comms protocol
before joining a game was changed for more security: it no
longer assumes a text-based telnet like protocol is being used.

This means that compatibility with old clients/servers was broken.

The next step is debugging and making sure it actually works.