Fixing up the low-level networking
There is no longer a need to a have a separate thread for
listening to incoming messages because this is taken care of
by libdeng2.

LogBuffer is now thread-safe as it gets automatically locked
when somebody accesses it.

Continued fixing up things so that the rather unusual setup of
having the QCoreApplication outside the main executable
will work.