LegacyNetwork now operates in blocking mode
It seems there is some fundamental issue with running
Qt's network code from inside a non-GUI event loop that
is created under the deng2 shared library. Until the
application itself is converted to use Qt (and link to deng1
as a static library), as a workaround LegacyNetwork now
blocks on socket writes and polls for incoming data.
These kludges can be removed once LegacyCore is out
of the picture.

Mac: The client is now successfully querying the server's
info with "net search" (N_LookForHosts).