Added: New console variable type CVT_URIPTR i.e., Uri
As Uri is now a fundamental data type in Doomsday - the console should
have built-in support and feature dynamic creation of variables of this
core type.

Mostly these will be handled identically to strings but with the added
benefit of centralised scheme syntax validation and error checking.
Note that this functionality has not yet been implemented - it can wait
until we derive Uri from Qt's QUrl.