Refactor: Revised use of texture "original indices"
As these cannot be eradicated for compatibility reasons I decided to
repurpose them for use as our namespace-unique identifiers and are now
used as the basis for texture URNs.

The Textures collection can now resolve URNs via the Textures_ResolveUri
interface and additionally provides methods for looking up textures by
their "unique id". The translation LUTs have accordingly been moved into
the collection. The collection also now provides a method for composing
a texture URN for a textureid_t

Refactor: Looking up a Material for a texture by it's unique id no longer
requires a Texture to be defined

Refactor: flat_t was removed as it is now redundant

Refactor: Redefined patchid_t as the Patch namespace's unique id and it
is no longer necessary for a Texture to be defined as a result of a name
being declaring on game side