Fixed: Loading Packages with ccmd load uses name pattern matching
When loading a Package using the console, the resource locator would
default to using the automatic name pattern matching logic even when
a specific extension is specified (e.g., loading a Zip when the user
requested a nonexistant Wad).

This resource locator logic has been optional for the specific case
outlined above to - if an extension is specified consider only those
resources which match.