Multiplayer: Fixed issues with psprite animations
Problems were fixed in the animation of the weapon psprite
on the client when picking up weapons/ammo/armor,
when first spawning on a level, when dying, and when
respawning back from the dead.

The game function P_MaybeChangeWeapon is now
executed by the client and never by the server, which
means the client's configuration affects the weapon
change rules.

Server protocol version number was incremented
because a new packet (GPT_MAYBE_CHANGE_WEAPON)
was added.