Created an empty dsfmod plugin for FMOD Ex
The FMOD audio plugin is not built by default. The build option
deng_fmod needs to be added to config_user.pri to enable it.

Also added deng_nosdlmixer to disable the built-in SDL_mixer
audio interface. When we remove SDL completely, it has to
be moved out of the engine (or discarded in favor of Qt-based
audio).