Win32: CD playback with FMOD
Testing with the Hexen CD. It seems there are some disprepancies with
track numbering, as even though the Hexen CD has a data track #1, the
audio tracks are also listed starting from #1.

The track numbers in the MAPINFO need to be adjusted after loading
so they'll conform to the original track numbering.