Hexen: Fixed earthquake view shake
In the original game the offset applied to the camera view during an
earthquake was calculated at 35Hz due to the fixed refresh rate. This
fixed limit has been removed however the view shake offset was still
calculated on each refresh frame, meaning it was updating much too
often and how much was dependent on frame rate.

Also, since build 386 a sharp tic's worth of latency was introduced
to the shake as the actual offsets were not applied until the next
time R_UpdateConsoleView was called.

The shake offset has now been moved alongside the calculations for
view bob, view height, etc..., in P_CalcHeight (this really needs a
new name...).