FMOD: Use reverb config calculated by engine
Now the environmental reverb parameters computed by the
engine are used in dsFMOD. They have many shortcomings,
though. The most noticeable is that small sectors inside
bigger ones should be considered to be part of the same
aural space -- currently there are rapid reverb changes e.g.
in stairs and doorways.

Also, the overall volume of the reverb for some environments
should be adjusted.