Started a new branch for dsFluidSynth
My initial approach is to have a copy of FluidSynth under
doomsday/external and compile with qmake into a static library.
The dsFluidSynth plugin would then link with this.

If this proves problematic, the fallback is to use FluidSynth's
own CMake build files.

It is probably easier to let FluidSynth handle its own audio output
(e.g., CoreAudio on the Mac) instead of trying to route the
output through Doomsday's Sfx interface to a suitable audio plugin.
However, it means that compiling FluidSynth will require additional
platform specific configuring.

Note the dependency to glib.