Hexen: Fixed bug in loading savegames
The savegame reader was reading one extra byte
at the end of the data of each moving floor. This was
causing the loader to believe that the saved data
is corrupt, when it in reality is not.

Fixes http://sourceforge.net/tracker/index.php?func=detail&aid=2862377&group_id=74815&atid=542099