Fixed: Loading of older saved games from Plutonia and TNT broken
Older saved game versions did not differentiate between versions of
Doom2 (i.e., Plutonia and TNT are marked as Doom2). A kludge has been
implemented during load whereby if we determine that the old save is
from *any* version of Doom2 - replace the marked gamemode with that
of the current game.