Builder: Clarifications and improvements for Pilot
Maintenance actions are now handled separately and only
allowed on the master (finish, new). Common tasks are no longer
allowed as the concept was invalid (who executes the task? what if
two clients start doing it at the same time?).

Added "purge" as a new task. The "deb_changes" task will only be
given to the "ubuntu" client. Removed the apt/debchanges platform
check in autobuild().