Do not set monochrome and upscaleAndSharpenPatches when defining fonts
These states are obsolete now that we texture variants however as far
as the game is concerned there is only one set of values per Patch.
Therefore the engine must not define these properties when building
composite fonts; define the patches by name but leave the rest of the
per-patch state set to their defaults (jDoom will also declare some
of the same patches when it adds the Arms ownership icons).

We instead manage the monochrome and upscaleAndSharpen state for font
patches within BitmapCompositeFont by utilizing the texture variant
features, defining a new texture specification for these.

This fixes the confusion over the monochrome|upscaleAndSharpen state
with Doom's weapon ownership patches.