Snowberry: Updated -game option to specify game mode
A new setting 'game-mode' was added to the profiles. 'game-mode'
results in the -game option on the command line. The old 'game'
setting still exists so that dependencies on it are valid. However,
it no longer has an effect on the command line.

In old profiles it is necessary to manually select the game mode.