Snowberry: Fixes for paths in command line options
The $DENGBASE expression was not working as expected.
The resulting path will now be an absolute path.

Also added $SBROOT to point to the folder where
snowberry.py is located.

On the Mac, -vdmap is now used to specify where
to load the PK3 files for the engine and the games.