Fixed: GL_LoadImageFromFile guessing of an image handler
When attempting to load a lump as a graphic, GL_LoadImageFromFile
was using the path rather than the name of the lump for basing its
name pattern guess. Presently, this path is always of zero-length
for LumpFiles.

Status message is now verbose2 and wording now better represents
what is happening here. The Patch format handler is not presently
executed from there, it needs to be refactored into place.