Added search path flag SPF_NO_DECEND
ResourceNamespace search paths are now defined including a set of
flags which govern how the path will be interpreted.

SPF_NO_DECEND means that only resources in the base of the search
target will be included when populating the namespace. Otherwise,
all resources below the base of the search target will be included.