InFine: Client can skip the script
Clients will now send a skip request to the server when an
event occurs during client-side script playback.

Modified the InFine packets to include the script Id so that
the client can request skip on the correct script.

Local id of the current client-side InFine script is available
via the DD_CURRENT_CLIENT_FINALE_ID integer.