Refactor: Moved path property out of LumpInfo
Now that derivatives of AbstractFile store all their lump records
in their own PathDirectory instances, it is no longer necessary
for LumpInfo to store a copy of the absolute path.

WadFile, ZipFile and LumpFile can now dynamically compose the full
vfs path to a lump they contain, should it be needed.

Todo: It should not be necessary to compose an absolute path when
instantiating LumpFile, or when pruning duplicates from the zip
LumpDirectory, or when resolving resource searches.

Todo: LumpInfo::name is now redundant, it can be refactored away.