Refactor: Moving MaterialArchive into the engine (from libcommon)
Savegames already have a facility for creating a dictionary of materials.
This is also needed for netgames, so the MaterialArchive class was moved
into the engine. I/O in the class was replaced with Reader and Writer.

Todo: Configure the libcommon savegame code to use the appropriate
Reader/Writer callbacks so that the data is read from/written to the
savegame file.