Busy mode: Partial progress display mode
The progress bar can now be configured to use only a subregion
of the indicator. Visually combines the multiple consecutive busy
tasks during the launch of the engine.

Adjusted indicator draw timer so that the position of the gear
continues from the same point where it left off in the previous
busy task.

Todo: Is it possible to get rid of the screen clears between the
busy tasks, to remove the remaining flicker?