Removed screen clear between "Starting up" and "Buffering" tasks
Now that multiple workers may be combined into what visually appears
to be single continues process with sub-tasks there is no need to
clear at this time.

The screen is still cleared after the resource location task, as we
not want that the last frame of the progress animation is captured
when doing a transition on subsequently entering busy mode again.