Multiplayer: Server sends its materials to clients
This will ensure that even though a different set of resources
has been loaded on the client, the materials will not be mixed up.

Uses MaterialArchive to pass the materials to the clients.

Removed obsolete code for client "lump translation", which was
originally intended to handle this issue but is now completely
outdated.