Client|Fixed: Player no longer gets stuck in monsters
The client will now make sure that no solid clmobjs are moved into
a local player. Since solid objects cannot enter into each other,
this situation should never occur.

The issue is caused by the discrepancy between server and client
game worlds due to latency: the server-side player position does
not match the client-side one, so monsters may move into a
position that seems valid on the server but in truth is not suitable
in the client-side world.