Builder|Homepage: Monitor /web, send updates to website
autobuild.py now has facilities for monitoring all files and
directories under the /web directory in the repository. When
files are added, changed, or removed, an update package
is delivered to the website.

The update package is a zip archive with the following contents:
- updated/*: all changed or new files, to be copied into /www;
  directory structure mirrors the one in /www
- removed.txt: listing of removed files and directories

The format of removed.txt is as follows:
    rm (filepath)
    rm (filepath2)
    …
    rmdir (dirpath)
    rmdir (dirpath2)
    …

I.e., on every line one removed path is listed. 'rm' denotes a
removed file while 'rmdir' means a removed directory. The
paths are relative to /www.

removed.txt is omitted from the update archive if no files or
directories were removed.