LumpDirectory|Refactor: Added flag LDF_UNIQUE_PATHS
LumpDirectory's instantiated with this flag will self-manage their
contents to ensure they contain only unique paths among the lumps
they catalog.

If a lump is published to the directory with a path that already
appears in the directory, the newest lump according to file load
order wins out.

When replacing a lump the older lump is pruned from the directory
and the newer lump is added to the end. Note that the path-search
order remains unchanged (except now, the path points to the newer
lump).