libcommon: Enhanced loadgame and savegame console commands
These two commands now also implement the logic for prompting the user
to confirm a load/save game request, previously used with the <quick>
save slot only.

Each command now accepts a new "confirm" argument for the purpose of
declaring confirmation without being prompted. Note that this logic is
additional to that of cvar "game-save-confirm", meaning the user can
leave this enabled yet still override it.

Added read-only cvar "game-save-last-slot" which as the name suggests,
is the logical slot number of the last used save slot. There is also
a new mnemonic <last>, which can be used as a symbolic reference which
is automatically resolved with both loadgame and savegame commands.

The console commands quickload and quicksave are essentially glorified
aliases now.