Refactor: Removed GL_SelectTexUnits()
This ancient function made the assumption that the only texture type
we use is GL_TEXTURE_2D, which, while presently correct, will not be
true for much longer. Furthermore, changing the active state of a GL
texture unit should not inherently result in the implicit enabling or
disabling of a specific texture type for that unit. Thus the logic of
this function is no longer suitable as a global utility routine.

The two remaining users of this function (the render lists and sprite
renderer modules) now implement their own versions of this function.