Adjusting player movement behavior
In the original Doom input was processed during the 35 Hz sharp
ticks because of ticcmd synchronization. However, since Doomsday's
core timer runs much faster, it is necessary to delay the processing
of certain events so that control behavior matches the previous
stable versions.

Added cvar "input-toggle-sharp" (default: 1) for controlling this
behavior. (If set to 0, all events as processed as quickly as possible.)