Fixed: Further mismanagement of GL texture state
Calling glBindTexture() outside the scope of the GL texture manager
is now forbidden. There are now two routines for binding a texture to
the currently active texture unit:

a) GL_BindTexture() - used with TextureVariant instances
b) GL_BindTextureUnmanaged() - used with "unmanaged" texture names

Calling GL_BindTextureUnmanaged() with the GL name of a managed
texture is now considered to be an error (note there is no mechanism
in place to prevent or detect this).

Todo: Refactor the remaining users of the unmanaged logic branches
to use managed textures instead.