Renderer: Render lists module is now aware of "managed" textures
The interface mechanics are little clunky due to the texture unit map
needing to support unmanaged textures also. Once unmanaged textures
are refactored away, the lists will no longer contain references to
GL textures; meaning the hashes and lists will not need to be cleared
following a texture reset.

This commit fixes the issue of the GL magnification filter for "world"
class textures not being updated when this option is changed.