Snowberry: Renamed chex profile to doomchex
The sort order for profiles is determined alphabetically based
on the identifier. Since we cannot rename the defaults profile,
renamed the chex profile so it appears below it.

A proper fix would be to modify the sort order of FormattedList
(and DropdownList for minimal mode), but that seems like a
disproportionate amount of work for such a small bug.