Unix|Fixed: PathMap did not consider the case of the Unix-style path
Absolute Unix style paths require special case handling due to the
root being nameless/zero-length.

Also implemented a bunch of unit tests for this class which are ran
automatically the first time this class is used in a _DEBUG build.

Hopefully this fixes the issue of some resource packs (e.g., DHTP)
not working on Unix.