Library|Unix: Check symlinks when loading plugins
Plugins to load are ultimately chosen based on the file extension.
It is also conceivable that symlinks are used to point to plugins at
some other location.

Now Doomsday does not try to load files that clearly are not shared
libraries when it loads its plugins.

Todo: Setup a dedicated subdirectory under the system lib directory
where Doomsday's plugins are installed.