Unix: Improved fix for MatchDirectory, font loading
PathDirectoryNode_MatchDirectory() was not working in the 64-bit gcc
build. Applied more robust checking for a zero-length fragment.

Also forced font file path to lower case, because the "Fonts" namespace
name was used as part of the resource path, and on Unix the file
system is case-sensitive => real font files could not be found.