Automap|Fixed: Default zoom level at map start
When initializing the automap, first update the map bounds
and only then set the default view scale.

Also added debug prints for monitoring the automap
geometry/metrics update.

Todo for later: Make automap view scale truly resolution
independent so it is not necessary to change it after the size
of the widget has changed.