Bindings|Infine|Fixed: Handling of events during finale script
The finale subsystem was being given events in a way that bypasses the binding
context stack. Consequently, some input events were being handled at the wrong
time or by the wrong responder. For instance, when the menu is open all events
should be handled by it, and a finale script in the background shouldn't be
getting any of them.

Added a new binding context "finale" for the sole purpose of responding to
finale events. Currently the context is always active, even when a finale
script is not running. The finale responder gets to make the decision whether
to eat an event or not. However, the "menu" context overrides "finale", so if
the menu is open it gets the events first.

In the future, finale script event triggers should be done via bindings in the
"finale" context. This naturally means that the "finale" context needs to be
activated and deactivated based on whether scripts are running.